Heisenbug fixed:)
This was caused by actual heap overflow. However, any attempts and debugging it led to the bug disappearing and it demonstrated in a really weird ways where it looked like anything from miscompile to stack corruption.

But the heap overflow was real. The optimization that only rendered the audio player frames if there was a change has the unfortunate sideffect that it bypassed the display blank wait and hence the main loop was super fast and every tick issued avr  status request, which could not be serviced in time so they kept growing. Of course any slowdown in the main loop meant the avr requests were serviced just in time and the bug disappeared.